fix(factura): cambiar kgCaja de int a float para soportar valores decimales (ej: 24.5)

+++ backend/apps/core/src/Service/Factura/Dto/FacturaDto.php
@@ -29,7 +29,7 @@ final class FacturaDto implements DtoRequestInterface
 
         public string $moneda = 'USD',
         public ?string $detalle = null,
-        public ?int $kgCaja = null,
+        public ?float $kgCaja = null,
         public ?string $unidadMedida = null,
         public ?int $cajas = null,
         public ?float $cantidad = null,
